move test into correct package and matcher into support-package
